requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
backports.zoneinfo==0.2.1; python_version < '3.9'
django-passwords==0.3.12
django==4.0.1
djangorestframework-simplejwt==5.0.0
djangorestframework==3.13.1
pyjwt==2.3.0; python_version >= '3.6'
pytz==2021.3
sqlparse==0.4.2; python_version >= '3.5'
